+++ load_test.py
@@ -1,16 +1,11 @@
-from settings import CLUSTER_URL, DB_NAME, COLLECTION_NAME
-from locust import User, task, between
-from faker import Faker
-from bson.decimal128 import Decimal128
-from decimal import Decimal
-from bson.codec_options import TypeCodec, TypeRegistry, CodecOptions
+from locust import task, between
 
-import time
+from mongo_user import MongoUser
+from settings import COLLECTION_NAME
+
 import pymongo
 import random
 
-# singleton Mongo client
-CLIENT = pymongo.MongoClient(CLUSTER_URL)
 
 # docs to insert per batch insert
 DOCS_PER_BATCH = 100
@@ -19,82 +14,16 @@ DOCS_PER_BATCH = 100
 NAMES_TO_CACHE = 1000
 
 
-class DecimalCodec(TypeCodec):
-    python_type = Decimal  # the Python type acted upon by this type codec
-    bson_type = Decimal128  # the BSON type acted upon by this type codec
-
-    def transform_python(self, value):
-        """
-        Function that transforms a custom type value into a type
-        that BSON can encode.
-        """
-        return Decimal128(value)
-
-    def transform_bson(self, value):
-        """
-        Function that transforms a vanilla BSON type value into our
-        custom type.
-        """
-        return value.to_decimal()
-
-
-class MongoUser(User):
+class MongoSampleUser(MongoUser):
     """
-    Sample mongodb workload generator
+    Generic sample mongodb workload generator
     """
     # no delays between operations
     wait_time = between(0.0, 0.0)
 
     def __init__(self, environment):
         super().__init__(environment)
-        self.db = None
-        self.collection, self.collection_secondary = None, None
-        self.faker = Faker()
         self.name_cache = []
-
-    def _process(self, name, func):
-        """
-        Run something in the locust context and time its execution
-        :param name: operation name to be used in stats output
-        :param func: what to run
-        """
-        start_time = time.time()
-        try:
-            func()
-        except Exception as e:
-            # output the error for debugging purposes
-            print(e)
-            total_time = int((time.time() - start_time) * 1000)
-            self.environment.events.request_failure.fire(
-                request_type='mongo', name=name, response_time=total_time, exception=e, response_length=0,
-            )
-        else:
-            total_time = int((time.time() - start_time) * 1000)
-            self.environment.events.request_success.fire(
-                request_type='mongo', name=name, response_time=total_time, response_length=1
-            )
-
-    def ensure_collection(self, coll_name, indexes):
-        """
-        Define the collection and its indexes
-        """
-        # prepare a codec for decimal values
-        decimal_codec = DecimalCodec()
-        type_registry = TypeRegistry([decimal_codec])
-        codec_options = CodecOptions(type_registry=type_registry)
-
-        # create the collection if not exists
-        if coll_name not in self.db.collection_names():
-            collection = self.db.create_collection(coll_name, codec_options=codec_options)
-        else:
-            collection = self.db.get_collection(coll_name, codec_options=codec_options)
-
-        # create the required indexes
-        if indexes:
-            collection.create_indexes(indexes)
-
-        # also return the second collection with readPreference=SECONDARY for analytical queries
-        return collection, self.db.get_collection(coll_name, read_preference=pymongo.read_preferences.Secondary())
 
     def generate_new_document(self):
         """
@@ -111,7 +40,7 @@ class MongoUser(User):
 
     def run_aggregation_pipeline(self):
         """
-        Run an aggregation pipeline on a analytical node
+        Run an aggregation pipeline on a secondary node
         """
         # count number of inhabitants per city
         group_by = {
@@ -139,7 +68,6 @@ class MongoUser(User):
         Executed every time a new test is started - place init code here
         """
         # prepare the collection
-        self.db = CLIENT[DB_NAME]
         index1 = pymongo.IndexModel([('first_name', pymongo.ASCENDING), ("last_name", pymongo.DESCENDING)],
                                     name="idx_first_last")
         self.collection, self.collection_secondary = self.ensure_collection(COLLECTION_NAME, [index1])
