+++ main.py
@@ -1,10 +1,10 @@
 from locust.env import Environment
-from load_test import MongoUser
+from load_test import MongoSampleUser
 
 
 if __name__ == '__main__':
     # setup Environment and Runner
-    env = Environment(user_classes=[MongoUser])
+    env = Environment(user_classes=[MongoSampleUser])
     env.create_local_runner()
     # start a WebUI instance
     env.create_web_ui("127.0.0.1", 8089)
